Add "Blocked Locations" to compare. (#2676)
* Cleaned up / centralized how we track disabled metrics.
* Added a "Blocked Locations" item to the "Show" list.
* Refactored the code substantially to make it easier to improve in the future.
* Tweaked sorting so that when diffs are equal (e.g. gained/lost metrics), we order higher-population locations before lower ones.
